refactor: cria pasta para labirinto e arquivo de parede com sua classe
