+++ src/labirinto/parede.py
@@ -1,0 +1,10 @@
+import pygame
+import random
+
+class Parede(pygame.sprite.Sprite):
+    """Classe para as paredes do labirinto"""
+    def __init__(self, x, y, lista_telhados):
+        super().__init__()
+        self.image = random.choice(lista_telhados)
+        self.rect = self.image.get_rect(topleft=(x, y))
+        self.mask = pygame.mask.from_surface(self.image) #Para não empacar nas paredes, o contato será somente com pixel direto
